+++ dev-tools/omdb/src/bin/omdb/db.rs
@@ -104,9 +104,6 @@ use nexus_db_model::Volume;
 use nexus_db_model::VolumeRepair;
 use nexus_db_model::VolumeResourceUsage;
 use nexus_db_model::VpcSubnet;
-use nexus_db_model::WebhookDelivery;
-use nexus_db_model::WebhookEventClass;
-use nexus_db_model::WebhookReceiver;
 use nexus_db_model::Zpool;
 use nexus_db_model::to_db_typed_uuid;
 use nexus_db_queries::context::OpContext;
@@ -145,8 +142,6 @@ use omicron_common::api::external::DataPageParams;
 use omicron_common::api::external::Generation;
 use omicron_common::api::external::InstanceState;
 use omicron_common::api::external::MacAddr;
-use omicron_common::api::external::NameOrId;
-use omicron_common::api::external::http_pagination::PaginatedBy;
 use omicron_uuid_kinds::CollectionUuid;
 use omicron_uuid_kinds::DatasetUuid;
 use omicron_uuid_kinds::DownstairsRegionUuid;
@@ -157,7 +152,6 @@ use omicron_uuid_kinds::PhysicalDiskUuid;
 use omicron_uuid_kinds::PropolisUuid;
 use omicron_uuid_kinds::SledUuid;
 use omicron_uuid_kinds::VolumeUuid;
-use omicron_uuid_kinds::WebhookEventUuid;
 use omicron_uuid_kinds::ZpoolUuid;
 use sled_agent_client::VolumeConstructionRequest;
 use std::borrow::Cow;
@@ -174,8 +168,11 @@ use std::sync::Arc;
 use strum::IntoEnumIterator;
 use tabled::Tabled;
 use uuid::Uuid;
+use webhook::WebhookArgs;
+use webhook::cmd_db_webhook;
 
 mod saga;
+mod webhook;
 
 const NO_ACTIVE_PROPOLIS_MSG: &str = "<no active Propolis>";
 const NOT_ON_SLED_MSG: &str = "<not on any sled>";
@@ -1158,94 +1155,6 @@ struct SetStorageBufferArgs {
     storage_buffer: i64,
 }
 
-#[derive(Debug, Args, Clone)]
-struct WebhookArgs {
-    #[command(subcommand)]
-    command: WebhookCommands,
-}
-
-#[derive(Debug, Subcommand, Clone)]
-enum WebhookCommands {
-    /// Get information on webhook receivers
-    #[clap(alias = "rx")]
-    Receiver {
-        #[command(subcommand)]
-        command: WebhookRxCommands,
-    },
-    /// Get information on webhook events
-    Event,
-    /// Get information on webhook delivieries
-    Delivery {
-        #[command(subcommand)]
-        command: WebhookDeliveryCommands,
-    },
-}
-
-#[derive(Debug, Subcommand, Clone)]
-enum WebhookRxCommands {
-    /// List webhook receivers
-    #[clap(alias = "ls")]
-    List(WebhookRxListArgs),
-
-    #[clap(alias = "show")]
-    Info(WebhookRxInfoArgs),
-}
-
-#[derive(Debug, Args, Clone)]
-struct WebhookRxInfoArgs {
-    receiver: NameOrId,
-}
-
-#[derive(Debug, Args, Clone)]
-struct WebhookRxListArgs {
-    #[clap(long, short = 'a')]
-    start_at: Option<Uuid>,
-}
-
-#[derive(Debug, Subcommand, Clone)]
-enum WebhookDeliveryCommands {
-    /// List webhook deliveries
-    #[clap(alias = "ls")]
-    List(WebhookDeliveryListArgs),
-
-    /// Show details on a webhook delivery, including its payload and attempt history.
-    #[clap(alias = "show")]
-    Info(WebhookDeliveryInfoArgs),
-}
-
-#[derive(Debug, Args, Clone)]
-struct WebhookDeliveryListArgs {
-    /// If present, show only deliveries to this receiver.
-    #[clap(long, short, alias = "rx")]
-    receiver: Option<NameOrId>,
-
-    /// If present, select only deliveries for the given event.
-    #[clap(long, short)]
-    event: Option<WebhookEventUuid>,
-
-    /// If present, select only deliveries in the provided state(s)
-    #[clap(long = "state", short)]
-    states: Vec<db::model::WebhookDeliveryState>,
-
-    /// If present, select only deliveries with the provided trigger(s)
-    #[clap(long = "trigger", short)]
-    triggers: Vec<db::model::WebhookDeliveryTrigger>,
-
-    /// Include only delivery entries created before this timestamp
-    #[clap(long, short)]
-    before: Option<DateTime<Utc>>,
-
-    /// Include only delivery entries created after this timestamp
-    #[clap(long, short)]
-    after: Option<DateTime<Utc>>,
-}
-
-#[derive(Debug, Args, Clone)]
-struct WebhookDeliveryInfoArgs {
-    /// The ID of the delivery to show.
-    delivery_id: Uuid,
-}
-
 impl DbArgs {
     /// Run a `omdb db` subcommand.
     ///
@@ -1557,7 +1466,7 @@ impl DbArgs {
                         command: OximeterCommands::ListProducers
                     }) => cmd_db_oximeter_list_producers(&datastore, fetch_opts).await,
 
-                    DbCommands::Webhook(WebhookArgs { command }) => cmd_db_webhook(&opctx, &datastore, &fetch_opts, command).await,
+                    DbCommands::Webhook(args) => cmd_db_webhook(&opctx, &datastore, &fetch_opts, &args).await,
                     DbCommands::Zpool(ZpoolArgs {
                         command: ZpoolCommands::List(args)
                     }) => cmd_db_zpool_list(&opctx, &datastore, &args).await,
@@ -8124,648 +8033,6 @@ async fn cmd_db_oximeter_list_producers(
     Ok(())
 }
 
-async fn cmd_db_webhook(
-    opctx: &OpContext,
-    datastore: &DataStore,
-    fetch_opts: &DbFetchOptions,
-    command: &WebhookCommands,
-) -> anyhow::Result<()> {
-    match command {
-        WebhookCommands::Receiver {
-            command: WebhookRxCommands::List(args),
-        } => cmd_db_webhook_rx_list(opctx, datastore, fetch_opts, args).await,
-        WebhookCommands::Receiver {
-            command: WebhookRxCommands::Info(args),
-        } => cmd_db_webhook_rx_info(datastore, fetch_opts, args).await,
-        WebhookCommands::Delivery {
-            command: WebhookDeliveryCommands::List(args),
-        } => cmd_db_webhook_delivery_list(datastore, fetch_opts, args).await,
-        WebhookCommands::Delivery {
-            command: WebhookDeliveryCommands::Info(args),
-        } => cmd_db_webhook_delivery_info(datastore, fetch_opts, args).await,
-        WebhookCommands::Event => {
-            Err(anyhow::anyhow!("not yet implemented, sorry!"))
-        }
-    }
-}
-
-async fn cmd_db_webhook_rx_list(
-    opctx: &OpContext,
-    datastore: &DataStore,
-    fetch_opts: &DbFetchOptions,
-    args: &WebhookRxListArgs,
-) -> anyhow::Result<()> {
-    let ctx = || {
-        if let Some(starting_at) = args.start_at {
-            format!("listing webhook receivers (starting at {starting_at}")
-        } else {
-            "listing webhook_receivers".to_string()
-        }
-    };
-    let pagparams = DataPageParams {
-        marker: args.start_at.as_ref(),
-        ..first_page(fetch_opts.fetch_limit)
-    };
-    let rxs = datastore
-        .webhook_rx_list(opctx, &PaginatedBy::Id(pagparams))
-        .await
-        .with_context(ctx)?;
-
-    check_limit(&rxs, fetch_opts.fetch_limit, ctx);
-
-    #[derive(Tabled)]
-    #[tabled(rename_all = "SCREAMING_SNAKE_CASE")]
-    struct RxRow {
-        id: Uuid,
-        #[tabled(display_with = "datetime_rfc3339_concise")]
-        created: chrono::DateTime<Utc>,
-        #[tabled(display_with = "datetime_rfc3339_concise")]
-        modified: chrono::DateTime<Utc>,
-        secrets: usize,
-        subscriptions: usize,
-        name: String,
-        endpoint: String,
-    }
-
-    let rows = rxs.into_iter().map(
-        |db::model::WebhookReceiverConfig { rx, secrets, subscriptions }| {
-            RxRow {
-                id: rx.id().into_untyped_uuid(),
-                name: rx.identity.name.to_string(),
-                created: rx.time_created(),
-                modified: rx.time_modified(),
-                secrets: secrets.len(),
-                subscriptions: subscriptions.len(),
-                endpoint: rx.endpoint,
-            }
-        },
-    );
-
-    let table = tabled::Table::new(rows)
-        .with(tabled::settings::Style::empty())
-        .with(tabled::settings::Padding::new(0, 1, 0, 0))
-        .to_string();
-    println!("{table}");
-
-    Ok(())
-}
-
-async fn cmd_db_webhook_rx_info(
-    datastore: &DataStore,
-    fetch_opts: &DbFetchOptions,
-    args: &WebhookRxInfoArgs,
-) -> anyhow::Result<()> {
-    use nexus_db_schema::schema::webhook_rx_event_glob::dsl as glob_dsl;
-    use nexus_db_schema::schema::webhook_rx_subscription::dsl as subscription_dsl;
-    use nexus_db_schema::schema::webhook_secret::dsl as secret_dsl;
-
-    let conn = datastore.pool_connection_for_tests().await?;
-    let rx = lookup_webhook_rx(datastore, &args.receiver)
-        .await
-        .with_context(|| format!("loading webhook receiver {}", args.receiver))?
-        .ok_or_else(|| {
-            anyhow::anyhow!("no webhook receiver {} exists", args.receiver)
-        })?;
-
-    const ID: &'static str = "ID";
-    const NAME: &'static str = "name";
-
-    const DESCRIPTION: &'static str = "description";
-    const CREATED: &'static str = "created at";
-    const DELETED: &'static str = "deleted at";
-    const MODIFIED: &'static str = "modified at";
-    const ENDPOINT: &'static str = "endpoint";
-    const GEN: &'static str = "generation";
-    const EXACT: &'static str = "exact subscriptions";
-    const GLOBS: &'static str = "glob subscriptions";
-    const GLOB_REGEX: &'static str = "  regex";
-    const GLOB_SCHEMA_VERSION: &'static str = "  schema version";
-    const GLOB_CREATED: &'static str = "  created at";
-    const GLOB_EXACT: &'static str = "  exact subscriptions";
-    const WIDTH: usize = const_max_len(&[
-        ID,
-        NAME,
-        DESCRIPTION,
-        CREATED,
-        DELETED,
-        MODIFIED,
-        ENDPOINT,
-        GEN,
-        EXACT,
-        GLOBS,
-        GLOB_REGEX,
-        GLOB_SCHEMA_VERSION,
-        GLOB_CREATED,
-        GLOB_EXACT,
-    ]);
-
-    let WebhookReceiver {
-        identity:
-            nexus_db_model::WebhookReceiverIdentity {
-                id,
-                name,
-                description,
-                time_created,
-                time_modified,
-                time_deleted,
-            },
-        endpoint,
-        secret_gen,
-        subscription_gen,
-    } = rx;
-
-    println!("\n{:=<80}", "== RECEIVER ");
-    println!("    {NAME:>WIDTH$}: {name}");
-    println!("    {ID:>WIDTH$}: {id}");
-    println!("    {DESCRIPTION:>WIDTH$}: {description}");
-    println!("    {ENDPOINT:>WIDTH$}: {endpoint}");
-    println!();
-    println!("    {CREATED:>WIDTH$}: {time_created}");
-    println!("    {MODIFIED:>WIDTH$}: {time_modified}");
-    if let Some(deleted) = time_deleted {
-        println!("    {DELETED:>WIDTH$}: {deleted}");
-    }
-
-    println!("\n{:=<80}", "== SECRETS ");
-    println!("    {GEN:>WIDTH$}: {}", secret_gen.0);
-
-    let query = secret_dsl::webhook_secret
-        .filter(secret_dsl::rx_id.eq(id.into_untyped_uuid()))
-        .select(db::model::WebhookSecret::as_select());
-    let secrets = if fetch_opts.include_deleted {
-        query.load_async(&*conn).await
-    } else {
-        query
-            .filter(secret_dsl::time_deleted.is_null())
-            .load_async(&*conn)
-            .await
-    };
-
-    match secrets {
-        Ok(secrets) => {
-            #[derive(Tabled)]
-            struct SecretRow {
-                id: Uuid,
-
-                #[tabled(display_with = "datetime_rfc3339_concise")]
-                created: chrono::DateTime<Utc>,
-
-                #[tabled(display_with = "datetime_opt_rfc3339_concise")]
-                deleted: Option<chrono::DateTime<Utc>>,
-            }
-            let rows = secrets.into_iter().map(
-                |db::model::WebhookSecret {
-                     identity:
-                         db::model::WebhookSecretIdentity {
-                             id,
-                             time_modified: _,
-                             time_created,
-                         },
-                     webhook_receiver_id: _,
-                     secret: _,
-                     time_deleted,
-                 }| SecretRow {
-                    id: id.into_untyped_uuid(),
-                    created: time_created,
-                    deleted: time_deleted,
-                },
-            );
-
-            let table = tabled::Table::new(rows)
-                .with(tabled::settings::Style::empty())
-                .with(tabled::settings::Padding::new(0, 1, 0, 0))
-                .to_string();
-            println!("{table}");
-        }
-        Err(e) => eprintln!("failed to list secrets: {e}"),
-    }
-
-    println!("\n{:=<80}", "== SUBSCRIPTIONS ");
-    println!("    {GEN:>WIDTH$}: {}", subscription_gen.0);
-
-    let exact = subscription_dsl::webhook_rx_subscription
-        .filter(subscription_dsl::rx_id.eq(id.into_untyped_uuid()))
-        .filter(subscription_dsl::glob.is_null())
-        .select(subscription_dsl::event_class)
-        .load_async::<WebhookEventClass>(&*conn)
-        .await;
-    match exact {
-        Ok(exact) => {
-            println!("    {EXACT:>WIDTH$}: {}", exact.len());
-            for event_class in exact {
-                println!("    - {event_class}");
-            }
-        }
-        Err(e) => {
-            eprintln!("failed to list exact subscriptions: {e}");
-        }
-    }
-
-    let globs = glob_dsl::webhook_rx_event_glob
-        .filter(glob_dsl::rx_id.eq(id.into_untyped_uuid()))
-        .select(db::model::WebhookRxEventGlob::as_select())
-        .load_async::<db::model::WebhookRxEventGlob>(&*conn)
-        .await;
-    match globs {
-        Ok(globs) => {
-            println!("    {GLOBS:>WIDTH$}: {}", globs.len());
-            for glob in globs {
-                let db::model::WebhookRxEventGlob {
-                    rx_id: _,
-                    glob: db::model::WebhookGlob { glob, regex },
-                    time_created,
-                    schema_version,
-                } = glob;
-                println!("    - {glob}");
-                println!("    {GLOB_CREATED:>WIDTH$}: {time_created}");
-                if let Some(v) = schema_version {
-                    println!("    {GLOB_SCHEMA_VERSION:>WIDTH$}: {v}")
-                } else {
-                    println!(
-                        "(i) {GLOB_SCHEMA_VERSION:>WIDTH$}: <not yet processed>",
-                    )
-                }
-
-                println!("    {GLOB_REGEX:>WIDTH$}: {regex}");
-                let exact = subscription_dsl::webhook_rx_subscription
-                    .filter(subscription_dsl::rx_id.eq(id.into_untyped_uuid()))
-                    .filter(subscription_dsl::glob.eq(glob))
-                    .select(subscription_dsl::event_class)
-                    .load_async::<WebhookEventClass>(&*conn)
-                    .await;
-                match exact {
-                    Ok(exact) => {
-                        println!("    {GLOB_EXACT:>WIDTH$}: {}", exact.len());
-                        for event_class in exact {
-                            println!("      - {event_class}")
-                        }
-                    }
-                    Err(e) => eprintln!(
-                        "failed to list exact subscriptions for glob: {e}"
-                    ),
-                }
-            }
-        }
-        Err(e) => {
-            eprintln!("failed to list glob subscriptions: {e}");
-        }
-    }
-
-    Ok(())
-}
-
-async fn cmd_db_webhook_delivery_list(
-    datastore: &DataStore,
-    fetch_opts: &DbFetchOptions,
-    args: &WebhookDeliveryListArgs,
-) -> anyhow::Result<()> {
-    use nexus_db_schema::schema::webhook_delivery::dsl as delivery_dsl;
-    let conn = datastore.pool_connection_for_tests().await?;
-    let mut query = delivery_dsl::webhook_delivery
-        .limit(fetch_opts.fetch_limit.get().into())
-        .order_by(delivery_dsl::time_created.desc())
-        .into_boxed();
-
-    if let (Some(before), Some(after)) = (args.before, args.after) {
-        anyhow::ensure!(
-            after < before,
-            "if both after and before are included, after must be earlier than before"
-        );
-    }
-
-    if let Some(before) = args.before {
-        query = query.filter(delivery_dsl::time_created.lt(before));
-    }
-
-    if let Some(after) = args.before {
-        query = query.filter(delivery_dsl::time_created.gt(after));
-    }
-
-    if let Some(ref receiver) = args.receiver {
-        let rx =
-            lookup_webhook_rx(datastore, receiver).await?.ok_or_else(|| {
-                anyhow::anyhow!("no webhook receiver {receiver} found")
-            })?;
-        query = query.filter(delivery_dsl::rx_id.eq(rx.identity.id));
-    }
-
-    if !args.states.is_empty() {
-        query = query.filter(delivery_dsl::state.eq_any(args.states.clone()));
-    }
-
-    if !args.triggers.is_empty() {
-        query = query
-            .filter(delivery_dsl::triggered_by.eq_any(args.triggers.clone()));
-    }
-
-    let ctx = || "listing webhook receivers";
-
-    let deliveries = query
-        .select(WebhookDelivery::as_select())
-        .load_async(&*conn)
-        .await
-        .with_context(ctx)?;
-
-    check_limit(&deliveries, fetch_opts.fetch_limit, ctx);
-
-    #[derive(Tabled)]
-    #[tabled(rename_all = "SCREAMING_SNAKE_CASE")]
-    struct DeliveryRow {
-        id: Uuid,
-        trigger: nexus_db_model::WebhookDeliveryTrigger,
-        state: nexus_db_model::WebhookDeliveryState,
-        attempts: u8,
-        #[tabled(display_with = "datetime_rfc3339_concise")]
-        time_created: DateTime<Utc>,
-        #[tabled(display_with = "datetime_opt_rfc3339_concise")]
-        time_completed: Option<DateTime<Utc>>,
-    }
-
-    #[derive(Tabled)]
-    #[tabled(rename_all = "SCREAMING_SNAKE_CASE")]
-    struct WithEventId<T: Tabled> {
-        #[tabled(inline)]
-        inner: T,
-        event_id: Uuid,
-    }
-
-    #[derive(Tabled)]
-    #[tabled(rename_all = "SCREAMING_SNAKE_CASE")]
-    struct WithRxId<T: Tabled> {
-        #[tabled(inline)]
-        inner: T,
-        receiver_id: Uuid,
-    }
-
-    impl From<&'_ WebhookDelivery> for DeliveryRow {
-        fn from(d: &WebhookDelivery) -> Self {
-            let WebhookDelivery {
-                id,
-                // event and receiver UUIDs are toggled on and off based on
-                // whether or not we are filtering by receiver and event, so
-                // ignore them here.
-                event_id: _,
-                rx_id: _,
-                attempts,
-                state,
-                time_created,
-                time_completed,
-                // ignore these as they are used for runtime coordination and
-                // aren't very useful for showing delivery history
-                deliverator_id: _,
-                time_leased: _,
-                triggered_by,
-            } = d;
-            Self {
-                id: id.into_untyped_uuid(),
-                trigger: *triggered_by,
-                state: *state,
-                attempts: attempts.0,
-                time_created: *time_created,
-                time_completed: *time_completed,
-            }
-        }
-    }
-
-    impl<'d, T> From<&'d WebhookDelivery> for WithEventId<T>
-    where
-        T: From<&'d WebhookDelivery> + Tabled,
-    {
-        fn from(d: &'d WebhookDelivery) -> Self {
-            Self { event_id: d.event_id.into_untyped_uuid(), inner: T::from(d) }
-        }
-    }
-
-    impl<'d, T> From<&'d WebhookDelivery> for WithRxId<T>
-    where
-        T: From<&'d WebhookDelivery> + Tabled,
-    {
-        fn from(d: &'d WebhookDelivery) -> Self {
-            Self { receiver_id: d.rx_id.into_untyped_uuid(), inner: T::from(d) }
-        }
-    }
-
-    let mut table = match (args.receiver.as_ref(), args.event) {
-        // Filtered by both receiver and event, so don't display either.
-        (Some(_), Some(_)) => {
-            tabled::Table::new(deliveries.iter().map(DeliveryRow::from))
-        }
-        // Filtered by neither receiver nor event, so include both.
-        (None, None) => tabled::Table::new(
-            deliveries.iter().map(WithRxId::<WithEventId<DeliveryRow>>::from),
-        ),
-        // Filtered by receiver ID only
-        (Some(_), None) => tabled::Table::new(
-            deliveries.iter().map(WithEventId::<DeliveryRow>::from),
-        ),
-        // Filtered by event ID only
-        (None, Some(_)) => tabled::Table::new(
-            deliveries.iter().map(WithRxId::<DeliveryRow>::from),
-        ),
-    };
-    table
-        .with(tabled::settings::Style::empty())
-        .with(tabled::settings::Padding::new(0, 1, 0, 0));
-
-    println!("{table}");
-    Ok(())
-}
-
-/// Helper function to look up a webhook receiver with the given name or ID
-async fn lookup_webhook_rx(
-    datastore: &DataStore,
-    name_or_id: &NameOrId,
-) -> anyhow::Result<Option<WebhookReceiver>> {
-    use nexus_db_schema::schema::webhook_receiver::dsl;
-
-    let conn = datastore.pool_connection_for_tests().await?;
-    match name_or_id {
-        NameOrId::Id(id) => {
-            dsl::webhook_receiver
-                .filter(dsl::id.eq(*id))
-                .limit(1)
-                .select(WebhookReceiver::as_select())
-                .get_result_async(&*conn)
-                .await
-        }
-        NameOrId::Name(ref name) => {
-            dsl::webhook_receiver
-                .filter(dsl::name.eq(name.to_string()))
-                .limit(1)
-                .select(WebhookReceiver::as_select())
-                .get_result_async(&*conn)
-                .await
-        }
-    }
-    .optional()
-    .with_context(|| format!("loading webhook_receiver {name_or_id}"))
-}
-
-async fn cmd_db_webhook_delivery_info(
-    datastore: &DataStore,
-    fetch_opts: &DbFetchOptions,
-    args: &WebhookDeliveryInfoArgs,
-) -> anyhow::Result<()> {
-    use db::model::WebhookDeliveryAttempt;
-    use nexus_db_schema::schema::webhook_delivery::dsl;
-    use nexus_db_schema::schema::webhook_delivery_attempt::dsl as attempt_dsl;
-
-    let WebhookDeliveryInfoArgs { delivery_id } = args;
-    let conn = datastore.pool_connection_for_tests().await?;
-    let delivery = dsl::webhook_delivery
-        .filter(dsl::id.eq(*delivery_id))
-        .limit(1)
-        .select(WebhookDelivery::as_select())
-        .get_result_async(&*conn)
-        .await
-        .optional()
-        .with_context(|| format!("loading webhook delivery {delivery_id}"))?
-        .ok_or_else(|| {
-            anyhow::anyhow!("no webhook delivery {delivery_id} exists")
-        })?;
-
-    const ID: &'static str = "ID";
-    const EVENT_ID: &'static str = "event ID";
-    const RECEIVER_ID: &'static str = "receiver ID";
-    const STATE: &'static str = "state";
-    const TRIGGER: &'static str = "triggered by";
-    const ATTEMPTS: &'static str = "attempts";
-    const TIME_CREATED: &'static str = "created at";
-    const TIME_COMPLETED: &'static str = "completed at";
-
-    const DELIVERATOR_ID: &'static str = "by Nexus";
-    const TIME_LEASED: &'static str = "leased at";
-
-    const WIDTH: usize = const_max_len(&[
-        ID,
-        EVENT_ID,
-        RECEIVER_ID,
-        TRIGGER,
-        STATE,
-        TIME_CREATED,
-        TIME_COMPLETED,
-        DELIVERATOR_ID,
-        TIME_LEASED,
-        ATTEMPTS,
-    ]);
-
-    let WebhookDelivery {
-        id,
-        event_id,
-        rx_id,
-        triggered_by,
-        attempts,
-        time_created,
-        time_completed,
-        state,
-        deliverator_id,
-        time_leased,
-    } = delivery;
-    println!("\n{:=<80}", "== DELIVERY ");
-    println!("    {ID:>WIDTH$}: {id}");
-    println!("    {EVENT_ID:>WIDTH$}: {event_id}");
-    println!("    {RECEIVER_ID:>WIDTH$}: {rx_id}");
-    println!("    {STATE:>WIDTH$}: {state}");
-    println!("    {TRIGGER:>WIDTH$}: {triggered_by}");
-    println!("    {TIME_CREATED:>WIDTH$}: {time_created}");
-    println!("    {ATTEMPTS}: {}", attempts.0);
-
-    if let Some(completed) = time_completed {
-        println!("\n{:=<80}", "== DELIVERY COMPLETED ");
-        println!("    {TIME_COMPLETED:>WIDTH$}: {completed}");
-        if let Some(leased) = time_leased {
-            println!("    {TIME_LEASED:>WIDTH$}: {leased}");
-        } else {
-            println!(
-                "/!\\ WEIRD: delivery is completed but has no start timestamp?"
-            );
-        }
-        if let Some(nexus) = deliverator_id {
-            println!("    {DELIVERATOR_ID:>WIDTH$}: {nexus}");
-        } else {
-            println!("/!\\ WEIRD: delivery is completed but has no Nexus ID?");
-        }
-    } else if let Some(leased) = time_leased {
-        println!("\n{:=<80}", "== DELIVERY IN PROGRESS ");
-        println!("    {TIME_LEASED:>WIDTH$}: {leased}");
-
-        if let Some(nexus) = deliverator_id {
-            println!("    {DELIVERATOR_ID:>WIDTH$}: {nexus}");
-        } else {
-            println!(
-                "/!\\ WEIRD: delivery is in progress but has no Nexus ID?"
-            );
-        }
-    } else if let Some(deliverator) = deliverator_id {
-        println!(
-            "/!\\ WEIRD: delivery is not completed or in progress but has \
-             Nexus ID {deliverator:?}"
-        );
-    }
-
-    // Okay, now go get attempts for this delivery.
-    let ctx = || format!("listing delivery attempts for {delivery_id}");
-    let attempts = attempt_dsl::webhook_delivery_attempt
-        .filter(attempt_dsl::delivery_id.eq(*delivery_id))
-        .order_by(attempt_dsl::attempt.desc())
-        .limit(fetch_opts.fetch_limit.get().into())
-        .select(WebhookDeliveryAttempt::as_select())
-        .load_async(&*conn)
-        .await
-        .with_context(ctx)?;
-
-    check_limit(&attempts, fetch_opts.fetch_limit, ctx);
-
-    if !attempts.is_empty() {
-        println!("\n{:=<80}", "== DELIVERY ATTEMPT HISTORY ");
-
-        #[derive(Tabled)]
-        #[tabled(rename_all = "SCREAMING_SNAKE_CASE")]
-        struct DeliveryAttemptRow {
-            id: Uuid,
-            #[tabled(rename = "#")]
-            attempt: u8,
-            #[tabled(display_with = "datetime_rfc3339_concise")]
-            time_created: DateTime<Utc>,
-            nexus_id: Uuid,
-            result: db::model::WebhookDeliveryAttemptResult,
-            #[tabled(display_with = "display_u16_opt")]
-            status: Option<u16>,
-            #[tabled(display_with = "display_time_delta_opt")]
-            duration: Option<chrono::TimeDelta>,
-        }
-
-        let rows = attempts.into_iter().map(
-            |WebhookDeliveryAttempt {
-                 id,
-                 delivery_id: _,
-                 rx_id: _,
-                 attempt,
-                 result,
-                 response_status,
-                 response_duration,
-                 time_created,
-                 deliverator_id,
-             }| DeliveryAttemptRow {
-                id: id.into_untyped_uuid(),
-                attempt: attempt.0,
-                time_created,
-                nexus_id: deliverator_id.into_untyped_uuid(),
-                result,
-                status: response_status.map(|u| u.into()),
-                duration: response_duration,
-            },
-        );
-        let mut table = tabled::Table::new(rows);
-        table
-            .with(tabled::settings::Style::empty())
-            .with(tabled::settings::Padding::new(0, 1, 0, 0));
-        println!("{table}");
-    }
-
-    Ok(())
-}
-
 // Format a `chrono::DateTime` in RFC3339 with milliseconds precision and using
 // `Z` rather than the UTC offset for UTC timestamps, to save a few characters
 // of line width in tabular output.
@@ -8878,12 +8145,4 @@ async fn cmd_db_zpool_set_storage_buffer(
     );
 
     Ok(())
-}
-
-fn display_time_delta_opt(t: &Option<chrono::TimeDelta>) -> String {
-    t.map(|t| t.to_string()).unwrap_or_else(|| "-".to_string())
-}
-
-fn display_u16_opt(u: &Option<u16>) -> String {
-    u.map(|u| u.to_string()).unwrap_or_else(|| "-".to_string())
 }
